% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimalAPD.R
\name{exampleAPDRE}
\alias{exampleAPDRE}
\title{Executable Example of APDRE Function}
\usage{
exampleAPDRE()
}
\value{
Provides message with example code for using the APDRE function with data included in the package
}
\description{
Example of APDRE function using data included in the package
}
