% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APDREcorr.R
\name{exampleAPDREcorr}
\alias{exampleAPDREcorr}
\title{Executable Example of APDREcorr Function}
\usage{
exampleAPDREcorr()
}
\value{
Prints message with example code for using the APDRE function using data included in the package
}
\description{
Example of APDREcorr function using data included in the package
}
