% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wolfexample}
\alias{wolfexample}
\title{Camera Trap Observations of African golden wolves}
\format{
Dataframe with 2 columns and 30 rows
Radians Time of observations, in radians from 0 to 2pi
SamplingPeriod variable identifying camera trap sampling period
}
\source{
\ Campbell L.A.D. 2017
}
\usage{
wolfexample
}
\description{
Example dataset with wolf observation times in radians and the sampling period during which each observation was recorded
}
\keyword{datasets}
